Refactor test utilities with pytest pythonpath config
- Add pythonpath = ["tests"] to pyproject.toml pytest config
- Create tests/helpers.py with shared utilities:
  - SUPPORTED_ARCHITECTURES, create_hardware, get_matrix_instructions,
    create_config_list
- Simplify conftest.py to fixtures only (hardware, generate_baseline)
- Update test imports to use helpers module
- Delete test_utils.py and __init__.py (no longer needed)
- Remove tests dir from PYTHONPATH in CMakeLists.txt

Users can now run pytest without manual PYTHONPATH setup, and test
utilities are not shipped with the installed package.
