+++ shared/origami/python/pyproject.toml
@@ -72,6 +72,7 @@ wheel.exclude = [
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
+pythonpath = ["tests"]
 python_files = ["test_*.py"]
 python_classes = ["Test*"]
 python_functions = ["test_*"]
