--- shared/origami/python/tests/__init__.py
@@ -1,5 +1,0 @@
-# Copyright Advanced Micro Devices, Inc., or its affiliates.
-# SPDX-License-Identifier: MIT
-
-"""Tests for origami Python bindings."""
-
