+++ shared/origami/python/tests/test_selector.py
@@ -11,7 +11,6 @@ import origami
 torch = pytest.importorskip("torch", reason="torch is required for OrigamiMatmulSelector tests.")
 
 from origami.selector import OrigamiMatmulSelector
-from .conftest import create_config_list
 
 
 class MockConfig:
